:bug:Fix: 함수 파일 분리 및 리뷰 보기 클릭 이벤트 수정 #29 #31
1. 리뷰와 관련된 함수 review.js에 옮기고 import 하는 방식으로
   바꾸었습니다.
2. 리뷰 보기 클릭 시 리뷰 나타남 & 추천바 사라짐. 다시 한 번 더 누르면
   리뷰 사라지고 추천바 나타남
