+++ templates/exhibition-detail.html
@@ -50,16 +50,17 @@
             <button class="show-reservation" id="reserveBtn">예약하기</button>
         </div>
         <!-- 이 부분에 리뷰 목록/ 동행글 목록 들어갈 자리 -->
+        <div class="rv-all-items-organizer" id="reviewList" style="display: none;"></div>
         <!-- 이 부분에 리뷰 목록/ 동행글 목록 들어갈 자리 -->
     </div>
-    <div class="recommend-organizer">
+    <div class="recommend-organizer" style="display: flex;">
         <h6 class="rec-intro">그 외 전시 추천</h6>
         <div class="recommend-organizer-small">
             <div class="one-by-one">
                 <div class="rec-img-box">
                     <a href="" target="" id="1-rec-img-anchor">
                         <img class="rec-img" src="../static/img/default-img.jpg" id="1-rec-img">
-                    </a>                    
+                    </a>
                 </div>
                 <span class="rec-title" id="1-rec-title">추천전시 제목</span>
             </div>
@@ -67,7 +68,7 @@
                 <div class="rec-img-box">
                     <a href="" target="" id="2-rec-img-anchor">
                         <img class="rec-img" src="../static/img/default-img.jpg" id="2-rec-img">
-                    </a>                    
+                    </a>
                 </div>
                 <span class="rec-title" id="2-rec-title">추천전시 제목</span>
             </div>
@@ -75,7 +76,7 @@
                 <div class="rec-img-box">
                     <a href="" target="" id="3-rec-img-anchor">
                         <img class="rec-img" src="../static/img/default-img.jpg" id="3-rec-img">
-                    </a>                    
+                    </a>
                 </div>
                 <span class="rec-title" id="3-rec-title">추천전시 제목</span>
             </div>
@@ -83,7 +84,7 @@
                 <div class="rec-img-box">
                     <a href="" target="" id="4-rec-img-anchor">
                         <img class="rec-img" src="../static/img/default-img.jpg" id="4-rec-img">
-                    </a>                    
+                    </a>
                 </div>
                 <span class="rec-title" id="4-rec-title">추천전시 제목</span>
             </div>
@@ -91,7 +92,7 @@
                 <div class="rec-img-box">
                     <a href="" target="" id="5-rec-img-anchor">
                         <img class="rec-img" src="../static/img/default-img.jpg" id="5-rec-img">
-                    </a>                    
+                    </a>
                 </div>
                 <span class="rec-title" id="5-rec-title">추천전시 제목</span>
             </div>
